News headlines app
This API allowed me to grab the headline. As a template, I followed this video on how to create an app. 

Python Flask Create News Web Application With News API
https://www.youtube.com/watch?v=I5syfw5zgnc

This video also helped me understand the new API better
https://www.youtube.com/watch?v=oUuIVJ96JsI&ab_channel=IzzyAnalytics

I went through a lot of issues with connecting my database to my application.py because I was trying to do it on my own but it was going well. This is where stack overflow helped me

https://stackoverflow.com/questions/55258575/insert-data-into-sqlite3-database-with-api

Even after learning how to connect my application and my SQLite database. It was not working with the todo method using post. So instead I just connected my application.py instead of trying to go todo post route.